+++ library/swift/mocks/MockEnvoyEngine.swift
@@ -3,7 +3,7 @@ import Foundation
 
 /// Mock implementation of `EnvoyEngine`. Used internally for testing the bridging layer & mocking.
 final class MockEnvoyEngine: NSObject {
-  init(runningCallback onEngineRunning: (() -> Void)? = nil, logger: ((String) -> Void)? = nil,
+  init(runningCallback onEngineRunning: ((Int) -> Void)? = nil, logger: ((String) -> Void)? = nil,
        eventTracker: (([String: String]) -> Void)? = nil, networkMonitoringMode: Int32 = 0) {}
 
   /// Closure called when `run(withConfig:)` is called.
